Skip predictions (lazy_)

+++ oof_predictions.py
@@ -64,6 +64,7 @@ def main():
     parser.add_argument("-w", "--workers", type=int, default=0)
     parser.add_argument("-d4", "--d4-tta", action="store_true")
     parser.add_argument("-hv", "--hv-tta", action="store_true")
+    parser.add_argument("-f", "--force-recompute", action="store_true")
 
     args = parser.parse_args()
 
@@ -74,7 +75,7 @@ def main():
 
     d4_tta = args.d4_tta
     hv_tta = args.hv_tta
-
+    force_recompute = args.force_recompute
     outputs = [OUTPUT_PRED_MODIFICATION_FLAG, OUTPUT_PRED_MODIFICATION_TYPE]
 
     for checkpoint_fname in checkpoint_fnames:
@@ -93,44 +94,50 @@ def main():
         fold = checkpoints[0]["checkpoint_data"]["cmd_args"]["fold"]
         _, valid_ds, _ = get_datasets(data_dir, fold=fold, features=required_features)
 
-        oof_predictions = compute_oof_predictions(model, valid_ds, batch_size=batch_size, workers=workers)
         oof_predictions_csv = fs.change_extension(checkpoint_fname, "_oof_predictions.csv")
-        oof_predictions.to_csv(oof_predictions_csv, index=False)
+        if not os.path.exists(oof_predictions_csv) or force_recompute:
+            oof_predictions = compute_oof_predictions(model, valid_ds, batch_size=batch_size, workers=workers)
+            oof_predictions.to_csv(oof_predictions_csv, index=False)
 
         if hv_tta:
-            tta_model = wrap_model_with_tta(model, "flip-hv", inputs=required_features, outputs=outputs).eval()
-            oof_predictions = compute_oof_predictions(tta_model, valid_ds, batch_size=batch_size, workers=workers)
             oof_predictions_csv = fs.change_extension(checkpoint_fname, "_oof_predictions_flip_hv_tta.csv")
-            oof_predictions.to_csv(oof_predictions_csv, index=False)
+            if not os.path.exists(oof_predictions_csv) or force_recompute:
+                tta_model = wrap_model_with_tta(model, "flip-hv", inputs=required_features, outputs=outputs).eval()
+                oof_predictions = compute_oof_predictions(tta_model, valid_ds, batch_size=batch_size, workers=workers)
+                oof_predictions.to_csv(oof_predictions_csv, index=False)
 
         if d4_tta:
-            tta_model = wrap_model_with_tta(model, "d4", inputs=required_features, outputs=outputs).eval()
-            oof_predictions = compute_oof_predictions(tta_model, valid_ds, batch_size=batch_size, workers=workers)
             oof_predictions_csv = fs.change_extension(checkpoint_fname, "_oof_predictions_d4_tta.csv")
-            oof_predictions.to_csv(oof_predictions_csv, index=False)
+            if not os.path.exists(oof_predictions_csv) or force_recompute:
+                tta_model = wrap_model_with_tta(model, "d4", inputs=required_features, outputs=outputs).eval()
+                oof_predictions = compute_oof_predictions(tta_model, valid_ds, batch_size=batch_size, workers=workers)
+                oof_predictions.to_csv(oof_predictions_csv, index=False)
 
         # Holdout
         holdout_ds = get_holdout(data_dir, features=required_features)
 
-        holdout_predictions = compute_oof_predictions(model, holdout_ds, batch_size=batch_size, workers=workers)
         holdout_predictions_csv = fs.change_extension(checkpoint_fname, "_holdout_predictions.csv")
-        holdout_predictions.to_csv(holdout_predictions_csv, index=False)
+        if not os.path.exists(holdout_predictions_csv) or force_recompute:
+            holdout_predictions = compute_oof_predictions(model, holdout_ds, batch_size=batch_size, workers=workers)
+            holdout_predictions.to_csv(holdout_predictions_csv, index=False)
 
         if hv_tta:
-            tta_model = wrap_model_with_tta(model, "flip-hv", inputs=required_features, outputs=outputs).eval()
-            holdout_predictions = compute_oof_predictions(
-                tta_model, holdout_ds, batch_size=batch_size, workers=workers
-            )
             holdout_predictions_csv = fs.change_extension(checkpoint_fname, "_holdout_predictions_flip_hv_tta.csv")
-            holdout_predictions.to_csv(holdout_predictions_csv, index=False)
+            if not os.path.exists(holdout_predictions_csv) or force_recompute:
+                tta_model = wrap_model_with_tta(model, "flip-hv", inputs=required_features, outputs=outputs).eval()
+                holdout_predictions = compute_oof_predictions(
+                    tta_model, holdout_ds, batch_size=batch_size, workers=workers
+                )
+                holdout_predictions.to_csv(holdout_predictions_csv, index=False)
 
         if d4_tta:
-            tta_model = wrap_model_with_tta(model, "d4", inputs=required_features, outputs=outputs).eval()
-            holdout_predictions = compute_oof_predictions(
-                tta_model, holdout_ds, batch_size=batch_size, workers=workers
-            )
             holdout_predictions_csv = fs.change_extension(checkpoint_fname, "_holdout_predictions_d4_tta.csv")
-            holdout_predictions.to_csv(holdout_predictions_csv, index=False)
+            if not os.path.exists(holdout_predictions_csv) or force_recompute:
+                tta_model = wrap_model_with_tta(model, "d4", inputs=required_features, outputs=outputs).eval()
+                holdout_predictions = compute_oof_predictions(
+                    tta_model, holdout_ds, batch_size=batch_size, workers=workers
+                )
+                holdout_predictions.to_csv(holdout_predictions_csv, index=False)
 
 
 if __name__ == "__main__":
